Improved table formatting
Previously, tables with much text in a row had had scroll bars to scroll through all the text instead of automatic line breaks.
This is due to an open issue in the sphinx read the docs theme (https://github.com/readthedocs/sphinx_rtd_theme/issues/1505).
A workaround has been implemented to fix this until the issue is resolved.

Additionally, grid tables have been replaced by list tables. This allows explicitly setting the width of the columns and makes it easier to change the tables in the future.