Rewrote, improved and updated packaging guide overview
The packaging guide overview has been rewritten, improved and updated.
A table that gives an overview of the different methods to create AppDirs and the corresponding AppImages has been added. It lists their advantages, disadvantages and differences.

Additionally, TODOs to remove the (outdated) rest of the overview and to make a section for each AppImage creation method have been added.