Improved, added and update signature information
The information on AppImage signatures has been improved and updated. Additional information about signatures has been added.

The signature page has been improved, with a more concise explanation of what's required and an overview over the different ways to embed a signature. Links to new information on specific AppImage creation tools have been added as well as a reference to the updating page. Explanations and the section order have also been improved.

A new signing section has been added to the pages of linuxdeploy, go-appimagetool and electron builder. Each contains new information on how to sign the AppImage with the respective tool.

Additionally:
- An embedding updating information section has been added to the go-appimagetool page.
- The updating page has been improved with links to the new go-appimage section and the signing section.
- Outdated information on the linuxdeploy page has been updated and links to more information as well as a warning have been added. Its embedding updating information section has also been improved.