Improved AppStream information
The AppStream page has been improved with a better introduction, a better explanation on why to use AppStream files, a new validating section and a new explanation on how to embed AppStream files with AppImage creation tools that create the AppDir.
An AppStream section has also been added to the linuxdeploy page, explaining how to embed an AppStream file using it.

Additionally, the packaging guide has been restructured. The updating, signature and AppStream pages have been moved from the optional features sub-section into the packaging guide section as the following pages are also optional and not more relevant.