Rewrote Software Overview page as AppImageKit
The software overview page has been rewritten and renamed to AppImageKit.
Previously, it had only contained AppImageKit and AppImageUpdate. But because AppImageUpdate is now extensively covered in other sections, it has been removed and a link to AppImageUpdate and other information has been added instead.
The introduction has been merged with the AppImageKit section introduction.
Descriptions, explanations and links in it have been improved.

The validate section has also been removed as validate has been moved into AppImageUpdate and isn't a part of AppImageKit anymore.

Additionally, a warning has been added which explains the current situation of the AppImageKit repository and its replacement with new repositories for the individual components.