Added other desktop integration tools
Information about the desktop integration tools Gear Lever, zap, Soar and AM has been added to the desktop integration page in the user guide.

Additionally, a warning has been added to the AppImageLauncher section that it doesn't work with current reference implementation AppImages anymore and to the appimaged section that the monitored directories are not configurable.
The project links have also been improved to not link to a specific version anymore.