Improved, updated and corrected testing page.
The testing page has been majorly improved, updated and corrected.

The introduction has been improved and duplicates in it have been removed. Everything in the introduction that only applies to live images has been moved into a corresponding section.

The testappimage part has been removed as the tool is outdated and doesn't work with modern AppImages. It has been replaced by an easy-to-understand explanation on how to manually test the AppImages with live images.

The appimage-testsuite part has also been removed as the tool and its supported distribution versions are outdated and as it contains some bugs, resulting in it not working out of the box. It has been replaced by a script inside the documentation, which is an improved version of the relevant parts of appimage-testsuite.
The following changes and improvements have been made:
- No dockerfiles are used. Instead, the container must exist locally or gets downloaded, making it future-proof. Important packages are still installed, but transparently in the script, and this is explained to the user.
- The docker command has been improved to use the X11 server socket to connect to the host system instead of an unnecessary TCP connection. (The original code didn't work anyways, as the code to get the IP didn't work on Linux and code to make the host X11 server listen on port 6000 was missing.)
- Non-executed und unnecessary code has been removed.
This way, several scripts could be merged into one smaller snippet which actually works without having to modify it. Explanatory comments have also been added to make the script easy to understand.

Additionally, the descriptions and explanations on the page have generally been improved.