Added a section about the alternative implementation
A section about the alternative (to the reference) implementation go-appimagetool has been added to the reference chapter. It explains the consequences of it being an alternative implementation and the differences between the implementations. It has been linked in relevant places.
The same information has also been added to the go-appimagetool page itself.

The appimagetool section in the reference implementation page has also been linked wherever it's mentioned to give more information on what it is. In some places where it makes sense, the information that it is (part of) the reference implementation has also been added.

Additionally, several minor improvements have been carried out:
- The signature page has been improved to explain when to use appimagetool.
- An additional reference implementation decision has been added.
- A warning that linuxdeploy doesn't use the new static runtime has been added to its page.
- All reference links have been made consistent by replacing some underscores with dashes.
- Phrasing has been improved.